Move spells_db into ProjectContext
It'll be exposed as turbo_batman, and contain all the context
properties. This way, we don't clutter the namespace, and have one place
to look for all the context properties.

Note that sizes didn't go into this, because:
a) It's used in too many places
b) It isn't really project specific - more to do with layout/ui

Resolves #5