Fix up PreparedSpells tab
It's now much more suitable for final inclusion.
It now has two separate delegates for edit mode and normal mode,
with a loader to switch between them. (It's much better than
managing a billion property changes through states.

It still needs work to add the auto-complete, and to actually write
stuff back to the model, but this is still progress.